requirements.txt
discord.py
discord.oauth2
python-dateutil
pytz